IDA operator release 24.0.1

+++ scripts/deployOperator.sh
@@ -8,12 +8,12 @@ source ${CUR_DIR}/scripts/helper/common.sh
 check_platform_version
 
 function show_help {
-    echo -e "\nUsage: deployOperator.sh -i operator_image -n namespace \n"
+    echo -e "\nUsage: deployOperator.sh -i operator_image \n"
     echo "Options:"
     echo "  -h  Display help"
     echo "  -i  Operator image name"
     echo "      For example: registry_url/ida-operator:version"
-    echo "  -n  The namespace to deploy Operator"
+    echo "  -c  The Operator scope, Cluster or Namespaced, the default is Namespaced"
     echo "  -s  Optional: Image pull secret, the default is empty"
     echo "      For example: ida-operator-secret"
 }
@@ -23,7 +23,7 @@ then
     show_help
     exit -1
 else
-    while getopts "h?i:n:s:" opt; do
+    while getopts "h?i:c:s:w:" opt; do
         case "$opt" in
         h|\?)
             show_help
@@ -31,9 +31,11 @@ else
             ;;
         i)  IMAGEREGISTRY=$OPTARG
             ;;
-        n)  NAMESPACE=$OPTARG
+        c)  SCOPE=$OPTARG
             ;;
         s)  SECRET=$OPTARG
+            ;;
+        w)  WATCH_NAMESPACE=$OPTARG
             ;;
         :)  echo "Invalid option: -$OPTARG requires an argument"
             show_help
@@ -44,12 +46,21 @@ else
 fi
 
 [ -f ./deployoperator.yaml ] && rm ./deployoperator.yaml
-cp ./descriptors/operator.yaml ./deployoperator.yaml
+[ -f ./operator-crd.yaml ] && rm ./operator-crd.yaml
+cp ./descriptors/operator-crd.yaml ./operator-crd.yaml
 
-[ -f ./cluster-role-binding.yaml ] && rm ./cluster-role-binding.yaml
-cp ./descriptors/cluster-role-binding.yaml ./cluster-role-binding.yaml
+if ([ ! -z ${SCOPE} ] && [[ ${SCOPE} == "Cluster" ]]) || [ ! -z ${WATCH_NAMESPACE} ]; then
+    cp ./descriptors/cluster/operator.yaml ./deployoperator.yaml
+    
+    [ -f ./cluster-role-binding.yaml ] && rm ./cluster-role-binding.yaml
+    cp ./descriptors/cluster/cluster-role-binding.yaml ./cluster-role-binding.yaml
 
-sed -e "s|<NAMESPACE>|$NAMESPACE|g" ./cluster-role-binding.yaml > ./cluster-role-binding_temp.yaml ;  mv ./cluster-role-binding_temp.yaml ./cluster-role-binding.yaml
+    NAMESPACE=$(oc config view --minify -o 'jsonpath={..namespace}')
+    sed -e "s|<NAMESPACE>|$NAMESPACE|g" ./cluster-role-binding.yaml > ./cluster-role-binding_temp.yaml ;  mv ./cluster-role-binding_temp.yaml ./cluster-role-binding.yaml
+else
+    cp ./descriptors/namespaced/operator.yaml ./deployoperator.yaml
+fi
+
 
 if [ ! -z ${IMAGEREGISTRY} ]; then
 # Change the location of the image
@@ -70,13 +81,34 @@ sed -e "s|imagePullSecrets:| |g" ./deployoperator.yaml > ./deployoperatorsav.yam
 sed -e "s|- name: <IMAGE_PULL_SECRET>| |g" ./deployoperator.yaml > ./deployoperatorsav.yaml ;  mv ./deployoperatorsav.yaml ./deployoperator.yaml
 fi
 
+if ([ ! -z ${SCOPE} ] && [[ ${SCOPE} == "Cluster" ]]) || [ ! -z ${WATCH_NAMESPACE} ]; then
+    oc apply -f ./descriptors/cluster/cluster-role.yaml
+    oc apply -f ./cluster-role-binding.yaml
+else
+    oc apply -f ./descriptors/namespaced/role.yaml
+    oc apply -f ./descriptors/namespaced/role-binding.yaml
+fi
 
-oc apply -f ./descriptors/operator-crd.yaml
+oc apply -f ./operator-crd.yaml
+oc apply -f ./descriptors/ida-operators-edit.yaml
 oc apply -f ./descriptors/service-account.yaml
-oc apply -f ./descriptors/role.yaml
-oc apply -f ./descriptors/role-binding.yaml
-oc apply -f ./descriptors/cluster-role.yaml
-oc apply -f ./cluster-role-binding.yaml
 
 oc apply -f ./deployoperator.yaml
+
+if [ ! -z ${WATCH_NAMESPACE} ]; then
+  oc patch deployment/ida-operator --type=json --patch '
+  [
+    { 
+      "op": "add",
+      "path": "/spec/template/spec/containers/0/env",
+      "value": [
+          {
+              "name": "WATCH_NAMESPACE",
+              "value": "'$WATCH_NAMESPACE'"
+          }
+       ]
+    }
+  ]
+  '
+fi
 echo -e "\033[32mAll descriptors have been successfully applied. Monitor the pod status with 'oc get pods -w'.\033[0m"
